fixed build error. now builds successfully
