+++ components/content/video-list.tsx
@@ -1,9 +1,21 @@
 import { Video } from "@/types/types";
+import clientPromise from "@/lib/mongodb";
 
 type VideoListProps = {
   videos: Video[];
   fallbackText: string;
 };
+
+export async function TutorialVideos() {
+  const client = await clientPromise;
+  const db = client.db("ShadowrunWeb");
+  const videos = await db
+    .collection<Video>("Videos")
+    .find({ isTutorial: "yes" })
+    .sort({ featuredOrder: 1 })
+    .toArray();
+  return videos as Video[];
+}
 
 export default async function VideoList({
   videos,
